FAISSRAGFastAPI/requirements.txt
# Web Framework and Server
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv

# Document Processing
pymupdf      # Imported as fitz
ebooklib        # For EPUB parsing
beautifulsoup4 # For cleaning HTML from EPUBs

# Vector Search and Machine Learning
faiss-cpu      # Use faiss-gpu if running on a Linux machine with CUDA
sentence-transformers
numpy
torch          # Required for device detection (CPU/CUDA)

# LLM and RAG Orchestration
langchain
langchain-openai
langchain-text-splitters
openai
fastapi 
uvicorn 
pymupdf 
faiss-cpu 
ebooklib 
beautifulsoup4 
sentence-transformers 
langchain-text-splitters 
langchain-openai 
python-dotenv